LayoutTests/svg/dom/getElementsById-in-disconnected-svg-expected.txt


PASS SVGSVGElement.prototype.getElementById must return the first element in tree order, within the 'svg' element's descendants, whose ID is elementId, or null if there is no such element
